Reconfigure webmock/VCR
