--- spec/fixtures/vcr_cassettes/NutritionFacade/Nutrition/_total_ingredients/can_produce_the_weight_of_all_ingredients_given.yml
@@ -1,60 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://api.api-ninjas.com/v1/nutrition?query=1%20lbs%20Chicken
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      X-Api-Key:
-      - X-Api-Key
-      User-Agent:
-      - Faraday v2.7.10
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '258'
-      Connection:
-      - keep-alive
-      Date:
-      - Sat, 29 Jul 2023 00:12:16 GMT
-      X-Amzn-Requestid:
-      - 21e6df1e-ec5d-486c-a977-517ff0adea95
-      Access-Control-Allow-Origin:
-      - "*"
-      Allow:
-      - GET, OPTIONS, POST
-      Access-Control-Allow-Headers:
-      - "*"
-      X-Amz-Apigw-Id:
-      - IzLnCHpUvHcF0Kg=
-      Access-Control-Allow-Methods:
-      - GET, OPTIONS, POST
-      X-Amzn-Trace-Id:
-      - Root=1-64c4595f-2f032b647cbc9cb02c2f30aa;Sampled=0;lineage=6ae1048e:0
-      X-Cache:
-      - Miss from cloudfront
-      Via:
-      - 1.1 12db2441e97bca59edef39f0800ba984.cloudfront.net (CloudFront)
-      X-Amz-Cf-Pop:
-      - DEN52-P1
-      X-Amz-Cf-Id:
-      - muBWjZSQWHgZeJwfO-ceOrUuI59BHLUZqlr95xGqMicPMSkyeQuepg==
-    body:
-      encoding: UTF-8
-      string: '[{"name": "chicken", "calories": 1009.9, "serving_size_g": 453.592,
-        "fat_total_g": 58.5, "fat_saturated_g": 16.6, "protein_g": 107.6, "sodium_mg":
-        327, "potassium_mg": 812, "cholesterol_mg": 417, "carbohydrates_total_g":
-        0.2, "fiber_g": 0.0, "sugar_g": 0.0}]'
-  recorded_at: Sat, 29 Jul 2023 00:12:16 GMT
-recorded_with: VCR 6.2.0
