--- spec/fixtures/vcr_cassettes/NutritionService/instance_methods/_ingredients_weight/can_take_in_a_string_of_ingredients_and_return_their_total_weight_in_grams.yml
@@ -1,60 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://api.api-ninjas.com/v1/nutrition?query=2%20lbs%20chicken
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      X-Api-Key:
-      - X-Api-Key
-      User-Agent:
-      - Faraday v2.7.10
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '260'
-      Connection:
-      - keep-alive
-      Date:
-      - Fri, 28 Jul 2023 23:33:36 GMT
-      X-Amzn-Requestid:
-      - 1be18ed8-26d5-4ee7-b44b-6392aac32964
-      Access-Control-Allow-Origin:
-      - "*"
-      Allow:
-      - GET, OPTIONS, POST
-      Access-Control-Allow-Headers:
-      - "*"
-      X-Amz-Apigw-Id:
-      - IzF8ZHw0vHcF3KQ=
-      Access-Control-Allow-Methods:
-      - GET, OPTIONS, POST
-      X-Amzn-Trace-Id:
-      - Root=1-64c4504f-451d6d066f509f0a52c7813e;Sampled=0;lineage=6ae1048e:0
-      X-Cache:
-      - Miss from cloudfront
-      Via:
-      - 1.1 85d3c7f5b91d4c8492ba27d65fa677d0.cloudfront.net (CloudFront)
-      X-Amz-Cf-Pop:
-      - DEN52-P1
-      X-Amz-Cf-Id:
-      - kSbVXVnZgzuT5idtr3j2WEfVMR9mRVMLaA3FopworANEAu9vgvODVA==
-    body:
-      encoding: UTF-8
-      string: '[{"name": "chicken", "calories": 2019.8, "serving_size_g": 907.184,
-        "fat_total_g": 117.0, "fat_saturated_g": 33.2, "protein_g": 215.3, "sodium_mg":
-        655, "potassium_mg": 1625, "cholesterol_mg": 835, "carbohydrates_total_g":
-        0.4, "fiber_g": 0.0, "sugar_g": 0.0}]'
-  recorded_at: Fri, 28 Jul 2023 23:33:36 GMT
-recorded_with: VCR 6.2.0
