--- spec/fixtures/vcr_cassettes/Users_API/Fetch_One_Merchant/happy_paths/can_get_one_merchant.yml
@@ -1,120 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://api.api-ninjas.com/v1/nutrition?query=2.0%20Chicken%20lbs%200.5%20Cheese%20lbs%202.0%20ground%20beef%20lbs%202.0%20eggs%20oz%20
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      X-Api-Key:
-      - X-Api-Key
-      User-Agent:
-      - Faraday v2.7.10
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '1065'
-      Connection:
-      - keep-alive
-      Date:
-      - Sat, 29 Jul 2023 18:19:14 GMT
-      X-Amzn-Requestid:
-      - d3f8283e-addf-48ab-abfa-e9c282ab2841
-      Access-Control-Allow-Origin:
-      - "*"
-      Allow:
-      - GET, OPTIONS, POST
-      Access-Control-Allow-Headers:
-      - "*"
-      X-Amz-Apigw-Id:
-      - I1q1XF0hvHcF2Mw=
-      Access-Control-Allow-Methods:
-      - GET, OPTIONS, POST
-      X-Amzn-Trace-Id:
-      - Root=1-64c55822-0ebd203f03e9822251de5014;Sampled=0;lineage=6ae1048e:0
-      X-Cache:
-      - Miss from cloudfront
-      Via:
-      - 1.1 96b078df4a5d96ad3cc52cfe9d984774.cloudfront.net (CloudFront)
-      X-Amz-Cf-Pop:
-      - DEN52-P1
-      X-Amz-Cf-Id:
-      - J7iaKGLc8ZYaNEBpz8OlahXkdg6T2naYKpeBvrpoN1i9ToTRsS115g==
-    body:
-      encoding: UTF-8
-      string: '[{"name": "chicken", "calories": 3784.9, "serving_size_g": 1700.0,
-        "fat_total_g": 219.3, "fat_saturated_g": 62.2, "protein_g": 403.4, "sodium_mg":
-        1228, "potassium_mg": 3045, "cholesterol_mg": 1566, "carbohydrates_total_g":
-        0.8, "fiber_g": 0.0, "sugar_g": 0.0}, {"name": "cheese", "calories": 8933.5,
-        "serving_size_g": 2267.96, "fat_total_g": 749.1, "fat_saturated_g": 429.2,
-        "protein_g": 515.9, "sodium_mg": 15008, "potassium_mg": 10431, "cholesterol_mg":
-        2271, "carbohydrates_total_g": 71.8, "fiber_g": 0.0, "sugar_g": 10.6}, {"name":
-        "ground beef", "calories": 25270.6, "serving_size_g": 9071.84, "fat_total_g":
-        1551.4, "fat_saturated_g": 587.8, "protein_g": 2470.2, "sodium_mg": 8342,
-        "potassium_mg": 20287, "cholesterol_mg": 8177, "carbohydrates_total_g": 0.0,
-        "fiber_g": 0.0, "sugar_g": 0.0}, {"name": "eggs", "calories": 13090.9, "serving_size_g":
-        9071.84, "fat_total_g": 849.8, "fat_saturated_g": 283.6, "protein_g": 1142.0,
-        "sodium_mg": 13034, "potassium_mg": 18147, "cholesterol_mg": 33888, "carbohydrates_total_g":
-        64.9, "fiber_g": 0.0, "sugar_g": 34.1}]'
-  recorded_at: Sat, 29 Jul 2023 18:19:14 GMT
-- request:
-    method: post
-    uri: https://beta4.api.climatiq.io/estimate
-    body:
-      encoding: UTF-8
-      string: '{"emission_factor":{"activity_id":"waste_type_food_waste-disposal_method_landfilled","source":"EPA","region":"US","year":2022,"source_lca_activity":"end_of_life","data_version":"2.2"},"parameters":{"weight":22111.64,"weight_unit":"g"}}'
-    headers:
-      Authorization:
-      - Authorization
-      'Content-Type: ':
-      - application/json
-      User-Agent:
-      - Faraday v2.7.10
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '691'
-      Content-Type:
-      - application/json
-      Access-Control-Allow-Origin:
-      - "*"
-      Cache-Control:
-      - private, s-maxage=0, max-age=600, must-revalidate
-      X-Content-Type-Options:
-      - nosniff
-      Content-Security-Policy:
-      - frame-ancestors 'none'
-      Strict-Transport-Security:
-      - max-age=3600
-      X-Served-By:
-      - cache-den8254-DEN
-      Date:
-      - Sat, 29 Jul 2023 18:19:13 GMT
-    body:
-      encoding: UTF-8
-      string: '{"co2e":14.136868307551117,"co2e_unit":"kg","co2e_calculation_method":"ar4","co2e_calculation_origin":"source","emission_factor":{"name":"Food
-        Waste - Landfilled","activity_id":"waste_type_food_waste-disposal_method_landfilled","id":"163972ef-d865-4bbe-ab3b-f8b2465cf825","access_type":"public","source":"EPA","source_dataset":"GHG
-        Emission Factors Hub","year":2022,"region":"US","category":"Food and Organic
-        Waste","source_lca_activity":"end_of_life","data_quality_flags":[]},"constituent_gases":{"co2e_total":14.136868307551117,"co2e_other":null,"co2":null,"ch4":null,"n2o":null},"activity_data":{"activity_value":0.024373910875088133,"activity_unit":"short
-        ton"},"audit_trail":"selector"}'
-  recorded_at: Sat, 29 Jul 2023 18:19:14 GMT
-recorded_with: VCR 6.2.0
